change AddPostActivity to NewPostFragment
